Added basic Store API to update entry metadata and headers on 304s.
The current implementation only updates StoreMap metadata, and only
for the shared memory cache. TODO: Update stored metadata and
HTTP headers. Support disk caches.

Introduced std::atomic_flag instead of std::atomic<bool>. It may
be more efficient (e.g., it is guaranteed to be lock-free), but
it is more difficult to use and might not be as portable.