Supply AccessLogEntry (ALE) to cache_peer_access ACL checks.
See previous commit (236b205) for motivation and disclaimers. This
change covers a yet another use case.

This case is more complex because of the significant call depth of the
peerAllowedToUse() check, which is a known design problem of its own.
Solving that problem would make most of these changes unnecessary for
ALE delivery. We did not solve it here because (a) we did not want to
risk associated code stability; (b) it would require a lot of out of
scope changes; and (c) these changes may be beneficial long-term as they
deliver potentially useful info to deeper peer selection code.