Fix %>ru logging consistency for large and small URLs
The solution for logging large URLs, introduced in ffdd638,
had a flaw: with it, large and small URLs were logged differently.
For example, whitespaces are stripped from small URLs, but are not
from large ones. With this fix, both %>ru and %ru take their values
after the identical 'cleanup' procedure.