store.cc "!empty || (empty && collapsingInitiator())" assertion
... in StoreEntry::collapsed().

Some empty entries in bag14 lack transients, resulting in false
collapsingInitiator(). The lack of transients is a known bug partially
fixed in master 4310f8b and awaiting a commit for bug 4579. And
collapsingInitiator() itself is buggy (fixed in master d2a6dcb).

Since StoreEntry::collapsed() is only used for statistics collection,
and the related code should be revamped to sync with the official fixes,
removing the assertion (essentially) is the best short-term solution.
Whether broken collapsingInitiator() should be a part of the collapsed()
condition is debatable, but the correct answer (if any) is unimportant.