Try implement delay pools for Http::Tunneler
The caller is responsible to set correct DelayID for the Http::Tunneler.
Please see the comments in the new code.

This patch also fixes the delay pools related code inside tunnel.cc, it
was completely broken (it is broken for squid-5.x and squid-4.x releases).

XXX: Probably we should take care of Delay pools inside Ssl::ServerBump
where we are building a fake Store entry  to store Bumping errors.