Check if the must-staple flag is set to server certificate
The initial implementation implements the OCSP-stapling but does not
check if the certificate issuer requires from the server to send the
OCSP status with the certificate.

This patch includes implementation for openssl-1.1.0 and latter releases
and for openssl-1.0.2 and earlier releases.

Openssl-1.1.0 and latter releases include API to parse the TLSFeature
extension. This extension include a list of flags with the supported
features. The status-request (aka OCSP must-staple) feature is one of
the reported features in this tls extension.

Openssl-1.0.2 and earlier does not include API to parse TLSFeature
extension. This patch adds a simple parser for TLSFeature, to check
if the request-status flag is set.
However it is not enough tested and maybe has problems.

Alternatively we can just check if the TLSFeature is present in the certificate:
currently the only feature which defined by RFCs for this extension is the
request-status.