The PeerConnector class should not close the server connection
... in the case of errors. It just calls callback with the apriopriate error
and the FwdState object will close the server connection if required.

Closing the server connection in PeerConnector class in some cases
may cause segfault or assertions if the PeerConnector::callback called
after closing the server connection.

Also the Security::PeerConnector::sslFinalized() must return false
in the case on error, which is the normal because the connection was
not finalised in this case.

XXX: probably we need to add some checks inside PeerConnector::callback method
and check if this method already called or the server connection is closed.

This patch should go to squid-v4.x releases too