Honor SWAPOUT_FAILED definition
Before this fix, when exceeding maximum_object_size, SMP Squid(rock)
threw an unexpected exception in StoreEntry::checkDisk().  This
happened because the StoreEntry had not been marked for removal by
the time storeSwapOutFileClosed() got control. On the other hand,
this did not happen in Non-SMP(ufs) because the StoreEntry had been
marked for removal before storeSwapOutFileClosed() got control
inside StoreEntry::checkCachable()).

Since SMP and non-SMP code paths are different, we can not rely (or
insist) that StoreEntry::checkCachable() will work in all cases.
Fortunately, to fix this, it is possible to just call releaseRequest() a
bit earlier in storeSwapOutFileClosed(). I checked that the neighboring
code should be OK with that.