Fixed prime/master terminology and polished some stale branch comments
"Prime" or "primary" is a better term for the started-first TCP track
than "master" because Squid already uses "master" for master
transactions. Also, "primary" is a better antonym for "spare" (the best
antonym for "master" would be "slave" or "helper", but master/slave does
not reflect the relationship between the two tracks well, and we already
use "helper" terminology for other things).

No functionality changes intended.