Fix multiple bugs in ConnStateData::mayTunnelUnsupportedProto
Using the pipeline.nrequests to match the first bumped connection is bad idea.
The pipeline.nrequests counts the requests ever seen in a connection so it
can not work in connections which is reused.
Instead use a counter which counts the parsed HTTP bumped requests.

For transparently intercepted connections we can use the pipeline.nrequests
to match the state while read the first request (pipeline.nrequests <= 1).