Removed slightly misleading/misplaced comment
The comment made me think that the worker-queues option has no effect in
non-SMP mode, which is incorrect. I also adjusted worker-queues
documentation to reduce that implication.

Besides handling giveEachWorkerItsOwnQueue cases, the commented code is
responsible for handing non-SMP cases, and the comment might mislead the
reader that the code is dedicated to true giveEachWorkerItsOwnQueue in
SMP cases. A reader may ask "How can we _share_ anything if the are no
other workers to share with?!"

Also, the effect of COMM_REUSEPORT is comm_open_listener() business; we
do not need to document it in the caller.