Simplified flags.connected_okay setting.
I was tempted to move flags.connected_okay setting to syncWithServerConn
but realized that we only set this flag after successful CONNECT (if
any). Further analysis showed that it is always (and only) set when
calling dispatch() so I moved it there. This change is slightly out of
scope but it may reduce the number of difficult questions/doubts like
whether flags.connected_okay should be set in syncWithServerConn().